# task-manager-backend/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "version": "1.0.0",
        "description": "API para la gestión de tareas"
    },
    "paths": {
        "/api/tasks": {
            "post": {
                "summary": "Crear una nueva tarea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean"
                                    }
                                },
                                "required": ["title"]
                            }
                        }
                    },
                    "responses": {
                        "201": {
                            "description": "Tarea creada"
                        },
                        "400": {
                            "description": "Error de validación"
                        }
                    }
                }
            },
            "get": {
                "summary": "Obtener todas las tareas",
                "responses": {
                    "200": {
                        "description": "Lista de tareas"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "summary": "Obtener tarea por ID",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID de la tarea",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Detalles de la tarea"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            },
            "put": {
                "summary": "Actualizar tarea",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID de la tarea",
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "Tarea actualizada"
                        },
                        "404": {
                            "description": "Tarea no encontrada"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Eliminar tarea",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID de la tarea",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            }
        }
    }
}